Move the colorways into directories by profile
