+++ src/components/colorways/dsa/galaxy_class.js
