+++ src/components/colorways/gmk/merlin.js
