+++ src/components/colorways/gmk/striker.js
