+++ src/components/colorways/gmk/ta_royal_alpha.js
