+++ src/components/colorways/gmk/wob.js
