+++ src/components/colorways/index.js
@@ -1,24 +1,24 @@
-import dsa_galaxy_class from './dsa_galaxy_class';
-import sa_carbon from './sa_carbon';
-import sa_danger_zone from './sa_danger_zone';
-import sa_jukebox from './sa_jukebox';
-import sa_modern_selectric from './sa_modern_selectric';
-import sa_nantucket_selectric from './sa_nantucket_selectric';
-import sa_oblivion_hagoromo from './sa_oblivion_hagoromo';
-import gmk_dolch from './gmk_dolch';
-import gmk_jamon from './gmk_jamon';
-import gmk_merlin from './gmk_merlin';
-import gmk_metaverse from './gmk_metaverse';
-import gmk_mizu from './gmk_mizu';
-import gmk_nautilus from './gmk_nautilus';
-import gmk_olivetti from './gmk_olivetti';
-import gmk_olivia from './gmk_olivia';
-import gmk_serika from './gmk_serika';
-import gmk_space_cadet from './gmk_space_cadet';
-import gmk_striker from './gmk_striker';
-import gmk_ta_royal_alpha from './gmk_ta_royal_alpha';
-import gmk_wob from './gmk_wob';
-import sa_vilebloom from './sa_vilebloom';
+import dsa_galaxy_class from './dsa/galaxy_class';
+import sa_carbon from './sa/carbon';
+import sa_danger_zone from './sa/danger_zone';
+import sa_jukebox from './sa/jukebox';
+import sa_modern_selectric from './sa/modern_selectric';
+import sa_nantucket_selectric from './sa/nantucket_selectric';
+import sa_oblivion_hagoromo from './sa/oblivion_hagoromo';
+import gmk_dolch from './gmk/dolch';
+import gmk_jamon from './gmk/jamon';
+import gmk_merlin from './gmk/merlin';
+import gmk_metaverse from './gmk/metaverse';
+import gmk_mizu from './gmk/mizu';
+import gmk_nautilus from './gmk/nautilus';
+import gmk_olivetti from './gmk/olivetti';
+import gmk_olivia from './gmk/olivia';
+import gmk_serika from './gmk/serika';
+import gmk_space_cadet from './gmk/space_cadet';
+import gmk_striker from './gmk/striker';
+import gmk_ta_royal_alpha from './gmk/ta_royal_alpha';
+import gmk_wob from './gmk/wob';
+import sa_vilebloom from './sa/vilebloom';
 import alphaCodes from './alphaCodes';
 import modCodes from './modCodes';
